unify time rendering components
