+++ frontend/src/app/bisq/bisq-block/bisq-block.component.html
@@ -24,7 +24,7 @@
                 <td>
                   &lrm;{{ block.time | date:'yyyy-MM-dd HH:mm' }}
                   <div class="lg-inline">
-                    <i class="symbol">(<app-time-since [time]="block.time / 1000" [fastRender]="true"></app-time-since>)</i>
+                    <i class="symbol">(<app-time kind="since" [time]="block.time / 1000" [fastRender]="true"></app-time>)</i>
                   </div>
                 </td>
               </tr>
