+++ frontend/src/app/bisq/bisq-transaction/bisq-transaction.component.html
@@ -35,7 +35,7 @@
                   <td>
                     &lrm;{{ bisqTx.time | date:'yyyy-MM-dd HH:mm' }}
                     <div class="lg-inline">
-                      <i class="symbol">(<app-time-since [time]="bisqTx.time / 1000" [fastRender]="true"></app-time-since>)</i>
+                      <i class="symbol">(<app-time kind="since" [time]="bisqTx.time / 1000" [fastRender]="true"></app-time>)</i>
                     </div>
                   </td>
                 </tr>
