+++ frontend/src/app/components/blocks-list/blocks-list.component.html
@@ -43,7 +43,7 @@
             &lrm;{{ block.timestamp * 1000 | date:'yyyy-MM-dd HH:mm' }}
           </td>
           <td class="mined" *ngIf="!widget" [class]="indexingAvailable ? '' : 'legacy'">
-            <app-time-since [time]="block.timestamp" [fastRender]="true"></app-time-since>
+            <app-time kind="since" [time]="block.timestamp" [fastRender]="true"></app-time>
           </td>
           <td *ngIf="auditAvailable" class="health text-right" [ngClass]="{'widget': widget, 'legacy': !indexingAvailable}">
             <a
