+++ frontend/src/app/components/difficulty-adjustments-table/difficulty-adjustments-table.component.html
@@ -13,7 +13,7 @@
         <td class="d-none d-md-block"><a [routerLink]="['/block' | relativeUrl, diffChange.height]">{{ diffChange.height
             }}</a></td>
         <td class="text-left">
-          <app-time-since [time]="diffChange.timestamp" [fastRender]="true"></app-time-since>
+          <app-time kind="since" [time]="diffChange.timestamp" [fastRender]="true"></app-time>
         </td>
         <td class="text-right">{{ diffChange.difficultyShorten }}</td>
         <td class="text-right" [style]="diffChange.change >= 0 ? 'color: #42B747' : 'color: #B74242'">
