+++ frontend/src/app/components/pool/pool.component.html
@@ -227,7 +227,7 @@
             &lrm;{{ block.timestamp * 1000 | date:'yyyy-MM-dd HH:mm' }}
           </td>
           <td class="mined">
-            <app-time-since [time]="block.timestamp" [fastRender]="true"></app-time-since>
+            <app-time kind="since" [time]="block.timestamp" [fastRender]="true"></app-time>
           </td>
           <td class="coinbase">
             <span class="badge badge-secondary scriptmessage longer">
