+++ frontend/src/app/components/transactions-list/transactions-list.component.html
@@ -6,7 +6,7 @@
     <div>
       <ng-template [ngIf]="tx.status.confirmed">&lrm;{{ tx.status.block_time * 1000 | date:'yyyy-MM-dd HH:mm' }}</ng-template>
       <ng-template [ngIf]="!tx.status.confirmed && tx.firstSeen">
-        <i><app-time-since [time]="tx.firstSeen" [fastRender]="true"></app-time-since></i>
+        <i><app-time kind="since" [time]="tx.firstSeen" [fastRender]="true"></app-time></i>
       </ng-template>
     </div>
   </div>
