+++ frontend/src/app/shared/shared.module.ts
@@ -25,8 +25,7 @@ import { BytesPipe } from './pipes/bytes-pipe/bytes.pipe';
 import { WuBytesPipe } from './pipes/bytes-pipe/wubytes.pipe';
 import { FiatCurrencyPipe } from './pipes/fiat-currency.pipe';
 import { BlockchainComponent } from '../components/blockchain/blockchain.component';
-import { TimeSinceComponent } from '../components/time-since/time-since.component';
-import { TimeUntilComponent } from '../components/time-until/time-until.component';
+import { TimeComponent } from '../components/time/time.component';
 import { ClipboardComponent } from '../components/clipboard/clipboard.component';
 import { QrcodeComponent } from '../components/qrcode/qrcode.component';
 import { FiatComponent } from '../fiat/fiat.component';
@@ -53,7 +52,6 @@ import { AddressComponent } from '../components/address/address.component';
 import { SearchFormComponent } from '../components/search-form/search-form.component';
 import { AddressLabelsComponent } from '../components/address-labels/address-labels.component';
 import { FooterComponent } from '../components/footer/footer.component';
-import { TimeSpanComponent } from '../components/time-span/time-span.component';
 import { AssetComponent } from '../components/asset/asset.component';
 import { AssetsComponent } from '../components/assets/assets.component';
 import { AssetsNavComponent } from '../components/assets/assets-nav/assets-nav.component';
@@ -88,8 +86,7 @@ import { GeolocationComponent } from '../shared/components/geolocation/geolocati
 @NgModule({
   declarations: [
     ClipboardComponent,
-    TimeSinceComponent,
-    TimeUntilComponent,
+    TimeComponent,
     QrcodeComponent,
     FiatComponent,
     TxFeaturesComponent,
@@ -129,7 +126,6 @@ import { GeolocationComponent } from '../shared/components/geolocation/geolocati
     TransactionsListComponent,
     AddressComponent,
     SearchFormComponent,
-    TimeSpanComponent,
     AddressLabelsComponent,
     FooterComponent,
     AssetComponent,
@@ -195,8 +191,7 @@ import { GeolocationComponent } from '../shared/components/geolocation/geolocati
     NgbCollapseModule,
     InfiniteScrollModule,
     FontAwesomeModule,
-    TimeSinceComponent,
-    TimeUntilComponent,
+    TimeComponent,
     ClipboardComponent,
     QrcodeComponent,
     FiatComponent,
@@ -232,7 +227,6 @@ import { GeolocationComponent } from '../shared/components/geolocation/geolocati
     TransactionsListComponent,
     AddressComponent,
     SearchFormComponent,
-    TimeSpanComponent,
     AddressLabelsComponent,
     FooterComponent,
     AssetComponent,
